+++ crates/holochain/src/conductor/conductor/tests.rs
@@ -473,7 +473,11 @@ async fn test_signing_error_during_genesis() {
         panic!("this should have been an error")
     };
 
-    assert_matches!(err, ConductorApiError::ConductorError(ConductorError::GenesisFailed { errors }) if errors.len() == 1);
+    if let ConductorApiError::ConductorError(inner) = err {
+        assert_matches!(*inner, ConductorError::GenesisFailed { errors } if errors.len() == 1);
+    } else {
+        panic!("this should have been an error too");
+    }
 }
 
 async fn make_signing_call(client: &mut WebsocketSender, cell: &SweetCell) -> AppResponse {
@@ -562,7 +566,6 @@ async fn test_signing_error_during_genesis_doesnt_bork_interfaces() {
     // TODO: match the errors more tightly
     assert_matches!(response, AdminResponse::Error(_));
     let response = make_signing_call(&mut app_client, &cell2).await;
-    dbg!(&response);
 
     assert_matches!(response, AppResponse::Error(_));
 
@@ -743,7 +746,11 @@ async fn test_installation_fails_if_genesis_self_check_is_invalid() {
         panic!("this should have been an error")
     };
 
-    assert_matches!(err, ConductorApiError::ConductorError(ConductorError::GenesisFailed { errors }) if errors.len() == 1);
+    if let ConductorApiError::ConductorError(inner) = err {
+        assert_matches!(*inner, ConductorError::GenesisFailed { errors } if errors.len() == 1);
+    } else {
+        panic!("this should have been an error too");
+    }
 }
 
 #[tokio::test(flavor = "multi_thread")]