+++ crates/holochain/src/core/ribosome/host_fn/sign_ephemeral.rs
@@ -46,22 +46,11 @@ pub mod wasm_test {
     use ::fixt::prelude::*;
     use hdk::prelude::*;
     use holochain_keystore::AgentPubKeyExt;
-    use holochain_state::host_fn_workspace::HostFnWorkspace;
     use holochain_wasm_test_utils::TestWasm;
 
     #[tokio::test(flavor = "multi_thread")]
     async fn ribosome_sign_ephemeral_test() {
-        let test_env = holochain_state::test_utils::test_cell_env();
-        let test_cache = holochain_state::test_utils::test_cache_env();
-        let env = test_env.env();
-        let author = fake_agent_pubkey_1();
-        crate::test_utils::fake_genesis(env.clone())
-            .await
-            .unwrap();
-        let workspace = HostFnWorkspace::new(env.clone(), test_cache.env(), author).await.unwrap();
-
-        let mut host_access = fixt!(ZomeCallHostAccess, Predictable);
-        host_access.workspace = workspace;
+        let host_access = fixt!(ZomeCallHostAccess, Predictable);
 
         let output: Vec<EphemeralSignatures> =
             crate::call_test_ribosome!(host_access, TestWasm::Sign, "sign_ephemeral", ()).unwrap();