+++ crates/holochain_zome_types/CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased](https://github.com/holochain/holochain/holochain_zome_types-v0.0.2-alpha.1...HEAD)
 
+### Added
+
+- Countersigning related functions and structs
+
 ## 0.0.4
 
 ## 0.0.3